Created sudoku::GridPotential class
The sudoku::Solver class is getting bloated...

This is the first step of an attempt to refactor the fork() workflow. I'm thinking that users can initialize a Solver with a GridPotential instead of letting the solver create their own. This lets the user observe what values are available, so that the forking can be done outside the class.

I will integrate this new class with Solver in a future commit.