Added sudoku::partition functions
The sudoku::partition functions are effectively "load-balancers" for the
parallel block count tracker.

In order to parallelize the block count tracker, we need to partition
the grid into CPUCount (e.g. 4) groups of cell positions. Ideally, when
we update the related cells during setCellValue/clearCellValue,
the related cells should divide evenly into the 4 partitions.

I could not find a single partition algorithm that works perfectly. The
algorithms "effectiveness" depends on the dimensions as well as the
number of partitions. To mitigate this, I provided several partition
functions that can we can measure with the NEW sudoku_partition_test
tool. I'm suspicious that finding the "optimal" partition for any given
dimensions and partition size is NP-complete.

I also added the Dimensions::getRelatedCells() method to obtain a
contiguous vector of cell positions related to the requested cell. These
related cells don't contain duplicates, and they don't contain the
requested cell position.